+++ Example_Plugin_PCSC/src/main/java/org/eclipse/keyple/example/plugin/pcsc/UseCase4_TransmitControl/Main_TransmitControl_Pcsc.java
@@ -116,7 +116,8 @@ public class Main_TransmitControl_Pcsc {
 
     // Prepare the selection by adding the created generic selection to the card selection scenario.
     cardSelectionManager.prepareSelection(
-        cardSelector, cardExtension.createGenericCardSelectionExtension());
+        cardSelector,
+        cardExtension.getGenericCardApiFactory().createGenericCardSelectionExtension());
 
     // Schedule the selection scenario, always notify card presence.
     cardSelectionManager.scheduleCardSelectionScenario(